# backend/docs/swagger.yaml
definitions:
  controller.GetComponentsResponse:
    properties:
      components:
        items:
          $ref: '#/definitions/entity.Component'
        type: array
    type: object
  controller.GetIngredientsResponse:
    properties:
      ingredients:
        items:
          $ref: '#/definitions/entity.Ingredient'
        type: array
    type: object
  entity.Component:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  entity.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  entity.Ingredient:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: ref
  title: ref
  version: "1.0"
paths:
  /components:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetComponentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 全栄養素取得 API
      tags:
      - Component
  /components/{componentId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 栄養素 ID
        in: path
        name: componentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetComponentsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 栄養素取得 API
      tags:
      - Component
  /ingredients:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetIngredientsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.ErrorResponse'
      summary: 全素材取得 API
      tags:
      - Ingredient
swagger: "2.0"